[feat/post_nickname]: ApiModule 분리
