+++ data/src/main/java/org/gdsc/data/di/NetworkModule.kt
@@ -7,9 +7,6 @@ import dagger.hilt.components.SingletonComponent
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import org.gdsc.data.network.AuthInterceptor
-import org.gdsc.data.network.LoginAPI
-import org.gdsc.data.network.RestaurantAPI
-import org.gdsc.data.network.UserAPI
 import org.gdsc.domain.repository.TokenManager
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -18,7 +15,6 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 class NetworkModule {
-
 
     @NormalClient
     @Provides
@@ -49,24 +45,6 @@ class NetworkModule {
                     .build()
             )
             .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideLoginApi(@NormalClient retrofit: Retrofit): LoginAPI {
-        return retrofit.create(LoginAPI::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideRestaurantApi(@NormalClient retrofit: Retrofit): RestaurantAPI {
-        return retrofit.create(RestaurantAPI::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideUserApi(@AuthClient retrofit: Retrofit): UserAPI {
-        return retrofit.create(UserAPI::class.java)
     }
 
     companion object {
